app-ai-service/pyproject.toml
[project]
name = "app-ai-service"
version = "0.1.0"
description = "Servicio de IA para combinar reviews y surveys y enviar analisis a report"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
